requirements.txt
# add packages you want to use here
requests
pyjwt